// frontend/src/types/Transporte/transporte.ts
export interface ReboqueDto {
  id: number;
  placa: string;
  marca?: string;
  modelo?: string;
  anoFabricacao?: number;
  renavam?: string;
  chassi?: string;
  tipoCarroceria?: string;
  uf?: string;
  capacidadeCarga?: number;
  tara?: number;
  rntrc?: string;
  proprietarioNome?: string;
  proprietarioCpfCnpj?: string;
  ativo: boolean;
  dataCriacao: string;
  dataAtualizacao?: string;
}

export interface ReboqueCreateDto {
  placa: string;
  marca?: string;
  modelo?: string;
  anoFabricacao?: number;
  renavam?: string;
  chassi?: string;
  tipoCarroceria: string;
  uf: string;
  capacidadeCarga?: number;
  tara: number;
  rntrc?: string;
  proprietarioNome?: string;
  proprietarioCpfCnpj?: string;
  ativo?: boolean;
}

export interface ReboqueUpdateDto extends ReboqueCreateDto {}

export interface ReboqueListDto {
  id: number;
  placa: string;
  marca?: string;
  modelo?: string;
  tipoCarroceria?: string;
  capacidadeCarga?: number;
  ativo: boolean;
}
// ...
export interface ReboqueFiltros {
  pageNumber?: number;
  pageSize?: number;
  busca?: string;
  incluirInativos?: boolean;
  tipoCarroceria?: string;
}